requirements.txt
Flask
Flask-SQLAlchemy
Flask-Security-Too
argon2-cffi